+++ app/controllers/api/v1/fibonacci_controller.rb
@@ -1,15 +1,23 @@
 class Api::V1::FibonacciController < Api::BaseController
 
   def create
-    if permitted_fibonacci[:n]
-      n = permitted_fibonacci[:n].to_i
-      start_time = Time.now
-      service = FibonacciAlgorithmService.call(n)
-      end_time = (Time.now - start_time) * 1000
-      @fibonacci = Struct.new(:id, :value, :result, :runtime).new(1, n, service.result, end_time)
-      render json: @fibonacci.to_json, status: :created
+    unless permitted_fibonacci[:n]
+      render json: { error: "Please set the param 'n'" }, status: :bad_request and return
+    end
+
+    n = permitted_fibonacci[:n].to_i
+    start_time = Time.now
+    service = FibonacciAlgorithmService.call(n)
+    end_time = (Time.now - start_time) * 1000
+    @fibonacci = Fibonacci.new(value: n,
+                               result: service.result,
+                               runtime: end_time)
+
+    if @fibonacci.save
+      render json: @fibonacci, status: :created
     else
-      render json: { error: "Please set the param 'n'" }, status: :bad_request
+      render json: { error: @fibonacci.errors.full_messages.to_sentence },
+             status: :unprocessable_entity and return
     end
   end