+++ db/migrate/20221021171427_create_fibonaccis.rb
@@ -1,0 +1,10 @@
+class CreateFibonaccis < ActiveRecord::Migration[7.0]
+  def change
+    create_table :fibonaccis do |t|
+      t.integer :value
+      t.integer :result, limit: 8
+      t.decimal :runtime
+      t.time :created_at
+    end
+  end
+end